+++ packages/core/lib/src/modes/column_mode.dart
@@ -1,0 +1,15 @@
+part of 'render_mode.dart';
+
+/// A render mode that builds HTML into a [Column] widget.
+class ColumnMode extends RenderMode {
+  /// Creates a column render mode.
+  const ColumnMode();
+
+  @override
+  Widget buildBodyWidget(
+    WidgetFactory wf,
+    BuildContext context,
+    List<Widget> children,
+  ) =>
+      wf.buildColumnWidget(context, children);
+}